feat: CQDG-843 refont studies page with global search
