--- src/views/Studies/components/PageContent/Studies/index.module.css
@@ -1,8 +1,0 @@
-.studyTabWrapper {
-  display: flex;
-  background-color: white;
-  padding: 20px;
-}
-.studyTabWrapper :global(.ant-pagination) {
-  margin-bottom: 0;
-}
