+++ src/views/Studies/components/PageContent/index.module.css
@@ -1,9 +1,12 @@
-.studyExplorePageContent {
+.pageContent {
   width: 100%;
-  display: flex;
-  position: relative;
 }
-
-.studyTitle {
-  margin-bottom: 0 !important;
+.pageContent .tableWrapper {
+  width: 100%;
+}
+.pageContent .title {
+  margin: 0;
+}
+.pageContent .search {
+  margin-bottom: 8px;
 }
