+++ src/views/Studies/components/SideBarFacet/index.module.css
@@ -15,13 +15,14 @@
 
 .collapseIcon {
   font-size: var(--sidebar-menu-item-icon-size);
-  color: var(--body-2);
+  color: var(--gray-7);
   display: flex;
   justify-content: flex-start;
 }
 
 .content {
   padding: var(--sidebar-content-panel-padding);
+  padding-top: 0;
 }
 
 .spin {
@@ -34,7 +35,7 @@
 
 .button {
   width: 21px;
-  color: var(--body-2);
+  color: var(--gray-7);
 }
 .button:hover {
   background-color: transparent;
