+++ src/views/Studies/components/SideBarFacet/index.tsx
@@ -11,7 +11,7 @@ import { INDEXES } from 'graphql/constants';
 import FilterList from 'components/uiKit/FilterList';
 import { FilterInfo } from 'components/uiKit/FilterList/types';
 
-import { STUDIES_EXPLORATION_QB_ID } from '../../utils/constant';
+import { STUDIES_REPO_QB_ID } from '../../utils/constant';
 
 import styles from './index.module.css';
 
@@ -47,7 +47,7 @@ const SideBarFacet = ({ className, extendedMappingResults, filterInfo }: ISideBa
             <FilterList
               key={INDEXES.STUDY}
               index={INDEXES.STUDY}
-              queryBuilderId={STUDIES_EXPLORATION_QB_ID}
+              queryBuilderId={STUDIES_REPO_QB_ID}
               extendedMappingResults={extendedMappingResults}
               filterInfo={filterInfo}
               filterGroupConfig={filterGroupConfig}
